Adding more tests and refactoring
